This is just a note.
When you merge the code, there was some conflicts in MainMenu.scene and it was dissapeard. There was no GameObject in MainMenu. Tried to revert but it was saying still we had conflicts. I just copy MainMenu.scene (Inventory work commit) in our history and paste it. I think your changes are gone. Couldn't find better way :D