Small bug fix.
When we start game, character's money was set to 0. It was because of a line in Character Controller.

It was trying to pull money data in PlayerPrefsController as "money". I had set string to "Money" earlier.

It is a simple example of how we can fall into error easily. I think we need a class called string literals. It won't only prevent this errors. It'll also help us when we need to translate our game to other languages.